requirements.txt
asgiref==3.2.10
autopep8==1.6.0
Django==3.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
pycodestyle==2.8.0
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
